% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_high_dim.r
\name{calc_tau}
\alias{calc_tau}
\title{Unconditional non-Gaussianity statistic (tau)}
\usage{
calc_tau(k, pa, ch)
}
\arguments{
\item{k}{moment degree}

\item{pa}{candidate parent column}

\item{ch}{candidate child column}
}
\value{
a single non-negative numeric statistic
}
\description{
Unconditional non-Gaussianity statistic (tau)
}
\keyword{internal}
