% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_camuv.r
\name{camuv_check_identified_causality}
\alias{camuv_check_identified_causality}
\title{Check that no pair within a variable set already has a parent relation}
\usage{
camuv_check_identified_causality(vars, P)
}
\arguments{
\item{vars}{integer vector of variable indices}

\item{P}{current parent list}
}
\value{
TRUE if no pair in \code{vars} has an identified causal relation yet
}
\description{
Faithful port of \code{CAMUV._check_identified_causality()}.
}
\keyword{internal}
