% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_camuv.r
\name{camuv_get_residual}
\alias{camuv_get_residual}
\title{Residual of one variable regressed on a set of explanatory variables}
\usage{
camuv_get_residual(X, explained_i, explanatory_ids, reg_fn)
}
\arguments{
\item{X}{data matrix}

\item{explained_i}{column index of the explained variable}

\item{explanatory_ids}{column indices of the explanatory variables
(possibly empty)}

\item{reg_fn}{regressor function from \code{\link[=resit_make_regressor]{resit_make_regressor()}}}
}
\value{
residual vector (length n)
}
\description{
Faithful port of \code{CAMUV._get_residual()}. With no explanatory variables
the column is returned as is.
}
\keyword{internal}
