% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.r
\name{check_glmnet_available}
\alias{check_glmnet_available}
\title{Check whether glmnet is available}
\usage{
check_glmnet_available(method)
}
\arguments{
\item{method}{name of the regression method that requires glmnet (for the
error message)}
}
\description{
If it is not available, raise an error indicating which regression method
required it.
}
\keyword{internal}
