% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.r
\name{check_lavaan_available}
\alias{check_lavaan_available}
\title{Check whether lavaan is available}
\usage{
check_lavaan_available()
}
\description{
Check whether lavaan is available
}
\keyword{internal}
