% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_with_imputation.r
\name{check_mice_available}
\alias{check_mice_available}
\title{Check that the 'mice' package is available}
\usage{
check_mice_available(context)
}
\arguments{
\item{context}{Short description used in the error message}
}
\description{
Check that the 'mice' package is available
}
\keyword{internal}
