% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{create_bootstrap_result}
\alias{create_bootstrap_result}
\title{Create a BootstrapResult}
\usage{
create_bootstrap_result(
  adjacency_matrices,
  total_effects,
  resampled_indices = NULL,
  causal_orders = NULL
)
}
\arguments{
\item{adjacency_matrices}{array (n_sampling x n_features x n_features)}

\item{total_effects}{array (n_sampling x n_features x n_features)}

\item{resampled_indices}{list of index vectors}

\item{causal_orders}{matrix (n_sampling x n_features). Each row is the causal
order of one sample.}
}
\value{
BootstrapResult (list with class attribute)
}
\description{
Create a BootstrapResult
}
\keyword{internal}
