% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_with_imputation.r
\name{default_cd_fit}
\alias{default_cd_fit}
\title{Default causal-discovery fit: joint estimation via lingam_multi_group()}
\usage{
default_cd_fit(X_list, prior_knowledge, apply_prior_knowledge_softly)
}
\arguments{
\item{X_list}{List of imputed datasets (one per repeat)}

\item{prior_knowledge}{Prior knowledge matrix (NULL allowed)}

\item{apply_prior_knowledge_softly}{Apply prior knowledge softly (logical)}
}
\value{
\verb{list(causal_order = <integer vector>, adjacency_matrices = <list>)}
}
\description{
Treats the imputed datasets as "groups" sharing a common causal order.
}
\keyword{internal}
