% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{diff_mutual_info}
\alias{diff_mutual_info}
\title{Difference of mutual information}
\usage{
diff_mutual_info(xi_std, xj_std, ri_j, rj_i)
}
\arguments{
\item{xi_std}{Standardized xi}

\item{xj_std}{Standardized xj}

\item{ri_j}{Residual of xi regressed on xj}

\item{rj_i}{Residual of xj regressed on xi}
}
\value{
Difference of mutual information
}
\description{
Difference of mutual information
}
\keyword{internal}
