% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{direct_lingam}
\alias{direct_lingam}
\title{Direct LiNGAM}
\usage{
direct_lingam(
  X,
  prior_knowledge = NULL,
  apply_prior_knowledge_softly = FALSE,
  measure = "pwling"
)
}
\arguments{
\item{X}{数値行列 (n_samples x n_features)}

\item{prior_knowledge}{事前知識行列 (n_features x n_features) または NULL
 0: x_i から x_j への有向パスなし
 1: x_i から x_j への有向パスあり
-1: 不明}

\item{apply_prior_knowledge_softly}{事前知識をソフトに適用するか (logical)}

\item{measure}{独立性の評価尺度 ("pwling" または "kernel")}
}
\value{
list(adjacency_matrix, causal_order)
}
\description{
Direct LiNGAM
}
