% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{entropy_approx}
\alias{entropy_approx}
\title{Maximum-entropy approximation of entropy}
\usage{
entropy_approx(u)
}
\arguments{
\item{u}{Input vector}
}
\value{
Approximate entropy value
}
\description{
Maximum-entropy approximation of entropy
}
\keyword{internal}
