% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.r
\name{estimate_adjacency_matrix}
\alias{estimate_adjacency_matrix}
\title{Estimate the adjacency matrix from a causal order}
\usage{
estimate_adjacency_matrix(
  X,
  causal_order,
  prior_knowledge = NULL,
  method = "adaptive_lasso",
  lambda = "BIC",
  init_method = "ols"
)
}
\arguments{
\item{X}{original data}

\item{causal_order}{causal order (vector of 1-based indices)}

\item{prior_knowledge}{prior-knowledge matrix (NULL allowed)}

\item{method}{regression method
"ols"           : ordinary least squares (default)
"lasso"         : LASSO regression (glmnet)
"adaptive_lasso": Adaptive LASSO (two-stage)
"ridge"         : Ridge regression (glmnet)}

\item{lambda}{LASSO penalty (NULL = automatic selection by cross-validation)
"lambda.min" : minimum prediction error
"lambda.1se" : 1SE rule (sparser)
"AIC"       : minimum AIC (no CV, fast)
"BIC"        : minimum BIC (no CV, fast, sparsest), default
"oracle"     : Adaptive LASSO only. Not usable with Ridge.}

\item{init_method}{estimation method for the initial weights of Adaptive LASSO
"ols"   : ordinary least squares (default)
"ridge" : Ridge regression}
}
\value{
adjacency matrix B (n_features x n_features)
}
\description{
Estimate the adjacency matrix from a causal order
}
\keyword{internal}
