% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_total_effect.r
\name{estimate_all_total_effects}
\alias{estimate_all_total_effects}
\title{Estimate the total causal effects between all variables at once}
\usage{
estimate_all_total_effects(
  X,
  lingam_result,
  method = "adaptive_lasso",
  lambda = "BIC",
  init_method = "ols"
)
}
\arguments{
\item{X}{Original data (n_samples x n_features)}

\item{lingam_result}{Return value of lingam_direct()}

\item{method}{Regression method ("ols", "lasso", "adaptive_lasso", "ridge")}

\item{lambda}{Lambda selection ("lambda.min", "lambda.1se", "AIC", "BIC")}

\item{init_method}{Method for estimating the initial weights of adaptive LASSO regression ("ols" or "ridge")}
}
\value{
Matrix of total causal effects (n_features x n_features).
\strong{Convention: \code{TE[i, j]} is the total causal effect from variable j to variable i (j -> i).}
Same index convention as the adjacency matrix \code{adjacency_matrix}. The sum of direct and indirect effects.
}
\description{
Estimate the total causal effects between all variables at once
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

model <- LiNGAM_sample_1000$data |>
  lingam_direct(reg_method = "ols")

LiNGAM_sample_1000$data |>
  estimate_all_total_effects(model)
}
