% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_total_effect.r
\name{estimate_total_effect}
\alias{estimate_total_effect}
\title{Estimate the total causal effect between two specified variables}
\usage{
estimate_total_effect(
  X,
  lingam_result,
  from_index,
  to_index,
  method = "adaptive_lasso",
  lambda = "BIC",
  init_method = "ols"
)
}
\arguments{
\item{X}{Original data (matrix or data.frame)}

\item{lingam_result}{Return value of lingam_direct()}

\item{from_index}{Cause variable (1-based index or variable name)}

\item{to_index}{Effect variable (1-based index or variable name)}

\item{method}{Regression method ("ols", "lasso", "adaptive_lasso", "ridge"). Default is adaptive_lasso}

\item{lambda}{Lambda selection ("lambda.min", "lambda.1se", "AIC", "BIC", "oracle"). Default is BIC}

\item{init_method}{Method for estimating the initial weights of adaptive LASSO regression ("ols" or "ridge")}
}
\value{
Estimated total causal effect
}
\description{
Estimate the total causal effect between two specified variables
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

model <- LiNGAM_sample_1000$data |>
  lingam_direct(reg_method = "ols")

LiNGAM_sample_1000$data |>
  estimate_total_effect(model, 4, 1)
}
