% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.r
\name{extract_adjacency_matrix}
\alias{extract_adjacency_matrix}
\title{Extract an adjacency matrix from a result object or matrix}
\usage{
extract_adjacency_matrix(adjacency_matrix)
}
\description{
Extract an adjacency matrix from a result object or matrix
}
\keyword{internal}
