% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{extract_partial_orders}
\alias{extract_partial_orders}
\title{Extract partial orders from prior knowledge}
\usage{
extract_partial_orders(pk)
}
\arguments{
\item{pk}{Prior knowledge matrix (NaN = unknown)}
}
\value{
matrix (n x 2), each row is a (from, to) partial order
}
\description{
Extract partial orders from prior knowledge
}
\keyword{internal}
