% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.r
\name{fit_adaptive_lasso}
\alias{fit_adaptive_lasso}
\title{Adaptive LASSO}
\usage{
fit_adaptive_lasso(
  y,
  Xp,
  lambda = "BIC",
  gamma_weight = 1,
  init_method = "ols"
)
}
\arguments{
\item{y}{response variable}

\item{Xp}{predictor matrix}

\item{lambda}{lambda selection method ("lambda.min", "lambda.1se", "AIC", "BIC", "oracle")}

\item{gamma_weight}{exponent of the weights}

\item{init_method}{estimation method for the initial weights ("ols" or "ridge")}
}
\value{
coefficient vector
}
\description{
Adaptive LASSO
}
\keyword{internal}
