% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.r
\name{fit_lasso}
\alias{fit_lasso}
\title{LASSO regression (lambda selection by information criterion or CV)}
\usage{
fit_lasso(y, Xp, lambda = "BIC")
}
\arguments{
\item{y}{response variable}

\item{Xp}{predictor matrix}

\item{lambda}{lambda selection method
"lambda.min" : minimum CV prediction error
"lambda.1se" : CV 1SE rule
"AIC"       : minimum AIC
"BIC"        : minimum BIC, default}
}
\value{
coefficient vector
}
\description{
LASSO regression (lambda selection by information criterion or CV)
}
\keyword{internal}
