% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{fit_ols}
\alias{fit_ols}
\title{OLS regression}
\usage{
fit_ols(y, Xp)
}
\arguments{
\item{y}{response variable (numeric vector)}

\item{Xp}{predictor matrix}
}
\value{
coefficient vector (excluding intercept)
}
\description{
OLS regression
}
\keyword{internal}
