% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.r
\name{fit_ridge_reg}
\alias{fit_ridge_reg}
\title{Ridge regression (lambda selection by information criterion or CV)}
\usage{
fit_ridge_reg(y, Xp, lambda = "BIC")
}
\arguments{
\item{y}{response variable}

\item{Xp}{predictor matrix}

\item{lambda}{lambda selection method
"lambda.min" : minimum CV prediction error
"lambda.1se" : CV 1SE rule
"AIC"       : minimum AIC
"BIC"        : minimum BIC, default
"oracle" is not usable (Adaptive LASSO only).}
}
\value{
coefficient vector
}
\description{
Ridge regression (lambda selection by information criterion or CV)
}
\keyword{internal}
