% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{get_paths}
\alias{get_paths}
\title{Get all paths between two specified variables and their bootstrap probabilities}
\usage{
get_paths(result, from_index, to_index, min_causal_effect = NULL)
}
\arguments{
\item{result}{BootstrapResult object}

\item{from_index}{Start index (1-based)}

\item{to_index}{End index (1-based)}

\item{min_causal_effect}{Minimum threshold for the causal effect (NULL = 0)}
}
\value{
data.frame (path, effect, probability)
}
\description{
Get all paths between two specified variables and their bootstrap probabilities
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data,
  n_sampling = 30L, reg_method = "ols", seed = 42
)
get_paths(bs_model, 1, 6)
}
