% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{get_total_causal_effects}
\alias{get_total_causal_effects}
\title{Get a list of total causal effects}
\usage{
get_total_causal_effects(result, min_causal_effect = NULL)
}
\arguments{
\item{result}{BootstrapResult object}

\item{min_causal_effect}{Minimum threshold for the causal effect (NULL = 0)}
}
\value{
data.frame (from, to, effect, probability)
}
\description{
Get a list of total causal effects
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data,
  n_sampling = 30L, reg_method = "ols", seed = 42
)

get_total_causal_effects(bs_model)
}
