% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{get_var_names}
\alias{get_var_names}
\title{Get variable names, falling back to x0, x1, ... when colnames is NULL}
\usage{
get_var_names(x)
}
\description{
Get variable names, falling back to x0, x1, ... when colnames is NULL
}
\keyword{internal}
