% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.r
\name{glance.LiMResult}
\alias{glance.LiMResult}
\title{Get a one-row summary of a LiMResult}
\usage{
\method{glance}{LiMResult}(x, ...)
}
\arguments{
\item{x}{The return value of \code{\link[=lingam_lim]{lingam_lim()}} (a \code{LiMResult} object)}

\item{...}{Unused}
}
\value{
A one-row data.frame(n_variables, n_edges, n_discrete, causal_order)
}
\description{
Like \code{\link[=glance.LingamResult]{glance.LingamResult()}}, with an additional \code{n_discrete} column giving
the number of discrete variables in the model.
}
\examples{
set.seed(1)
dat <- generate_lim_sample(n = 300)
model <- lingam_lim(dat$data, is_continuous = dat$is_continuous)
glance(model)
}
