% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsic.r
\name{hsic_gram_matrix}
\alias{hsic_gram_matrix}
\title{Gaussian Gram matrix and its double-centered version}
\usage{
hsic_gram_matrix(x, width)
}
\arguments{
\item{x}{numeric vector}

\item{width}{kernel width from \code{\link[=hsic_kernel_width]{hsic_kernel_width()}}}
}
\value{
list(K = Gram matrix, Kc = centered Gram matrix)
}
\description{
Gaussian Gram matrix and its double-centered version
}
\keyword{internal}
