% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{ic_glmnet}
\alias{ic_glmnet}
\title{Select lambda by information criterion}
\usage{
ic_glmnet(glmnet_model)
}
\arguments{
\item{glmnet_model}{a glmnet model object}
}
\value{
list with lambda_AICc_best, lambda_BIC_best, ic_table
}
\description{
Select lambda by information criterion
}
\keyword{internal}
