% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_varma_total_effect.r
\name{joined_varma_matrix}
\alias{joined_varma_matrix}
\title{Join psi and omega arrays into a single wide matrix}
\usage{
joined_varma_matrix(psis, omegas)
}
\arguments{
\item{psis}{array (1 + p, m, m)}

\item{omegas}{array (q, m, m)}
}
\value{
matrix (m x m(1 + p + q))
}
\description{
Returns \code{cbind(psi_0, ..., psi_p, omega_1, ..., omega_q)} with shape
(m x m(1 + p + q)), keeping the \verb{[i, j]} = j -> i convention in each block.
}
\keyword{internal}
