% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{kernel_mi_core}
\alias{kernel_mi_core}
\title{カーネル法の相互情報量：本体}
\usage{
kernel_mi_core(E1, x2, kappa, sigma)
}
\arguments{
\item{E1}{\code{kernel_mi_prepare()} で前計算した変数1側の行列}

\item{x2}{変数2のベクトル}

\item{kappa}{正則化パラメータ}

\item{sigma}{ガウスカーネルの幅}
}
\value{
相互情報量
}
\description{
求める量は 2n x 2n 行列の logdet の差だが、ブロック構造と Schur 補行列により
n x n の Cholesky 分解だけで等価に計算できる：
\verb{MI = -1/2 * (logdet(tmp2^2 - K2 K1 tmp1^-2 K1 K2) - logdet(tmp2^2))}
}
\keyword{internal}
