% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{kurtosis_pop}
\alias{kurtosis_pop}
\title{尖度（n割り、正規分布で 0 になる超過尖度）}
\usage{
kurtosis_pop(x)
}
\arguments{
\item{x}{数値ベクトル}
}
\description{
尖度（n割り、正規分布で 0 になる超過尖度）
}
\keyword{internal}
