% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.r
\name{lavaan_fit_measures_to_df}
\alias{lavaan_fit_measures_to_df}
\title{Map lavaan fitMeasures() to the semopy-style column set}
\usage{
lavaan_fit_measures_to_df(fit)
}
\description{
Map lavaan fitMeasures() to the semopy-style column set
}
\keyword{internal}
