% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_lim.r
\name{lim_flatten_rowmajor}
\alias{lim_flatten_rowmajor}
\title{Flatten a matrix in row-major (numpy "C") order}
\usage{
lim_flatten_rowmajor(M)
}
\description{
Flatten a matrix in row-major (numpy "C") order
}
\keyword{internal}
