% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_lim.r
\name{lim_matpow}
\alias{lim_matpow}
\title{Matrix power by binary exponentiation}
\usage{
lim_matpow(M, k)
}
\arguments{
\item{M}{square matrix}

\item{k}{positive integer exponent}
}
\value{
\code{M} raised to the \code{k}-th power
}
\description{
Matrix power by binary exponentiation
}
\keyword{internal}
