% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_lim.r
\name{lim_obj_grad}
\alias{lim_obj_grad}
\title{Value and gradient of the augmented-Lagrangian objective for the doubled
parameter vector w (length 2\emph{d}d)}
\usage{
lim_obj_grad(w, d, X, con, W_dis_mask, W_con_mask, rho, alpha, lambda1)
}
\description{
Value and gradient of the augmented-Lagrangian objective for the doubled
parameter vector w (length 2\emph{d}d)
}
\keyword{internal}
