% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_lim.r
\name{lim_topological_order}
\alias{lim_topological_order}
\title{Topological order (Kahn's algorithm) from a lingamr-convention adjacency
matrix B (\code{B[i, j]} = j -> i, i.e. row i's nonzero columns are its parents)}
\usage{
lim_topological_order(B)
}
\description{
Topological order (Kahn's algorithm) from a lingamr-convention adjacency
matrix B (\code{B[i, j]} = j -> i, i.e. row i's nonzero columns are its parents)
}
\keyword{internal}
