% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_parce.r
\name{parce_residual}
\alias{parce_residual}
\title{Residual of \code{X[, j]} regressed on \code{X[, xi_index]} via the pseudo-inverse
of the covariance matrix}
\usage{
parce_residual(X, xi_index, j, Cov)
}
\arguments{
\item{X}{data matrix}

\item{xi_index}{explanatory-variable indices (may be empty)}

\item{j}{target variable index}

\item{Cov}{precomputed \code{stats::cov(X)}, since \code{X} is invariant across all
calls within a single \code{\link[=parce_search_causal_order]{parce_search_causal_order()}} search}
}
\value{
residual vector
}
\description{
Residual of \code{X[, j]} regressed on \code{X[, xi_index]} via the pseudo-inverse
of the covariance matrix
}
\keyword{internal}
