% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{print.BootstrapResult}
\alias{print.BootstrapResult}
\title{Display the contents of a BootstrapResult}
\usage{
\method{print}{BootstrapResult}(x, ...)
}
\arguments{
\item{x}{BootstrapResult object}

\item{...}{Additional arguments (for S3 method compatibility)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Display the contents of a BootstrapResult
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data,
  n_sampling = 30L, reg_method = "ols", seed = 42
)

print(bs_model)
}
