% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{print.lingam_normality_test}
\alias{print.lingam_normality_test}
\title{Print method for lingam_normality_test}
\usage{
\method{print}{lingam_normality_test}(x, ...)
}
\arguments{
\item{x}{lingam_normality_test object}

\item{...}{additional arguments}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for lingam_normality_test
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()
result <- lingam_direct(LiNGAM_sample_1000$data, reg_method = "ols")
print(test_residual_normality(LiNGAM_sample_1000$data, result))
}
