% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lingam.r
\name{print.lingam_summary}
\alias{print.lingam_summary}
\title{print method for lingam_summary}
\usage{
\method{print}{lingam_summary}(x, ...)
}
\arguments{
\item{x}{A \code{lingam_summary} object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
print method for lingam_summary
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()
model <- lingam_direct(LiNGAM_sample_1000$data, reg_method = "ols")
print(summary_lingam(LiNGAM_sample_1000$data, model))
}
