% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{print.LingamResult}
\alias{print.LingamResult}
\title{Print method for LingamResult}
\usage{
\method{print}{LingamResult}(x, digits = 3, ...)
}
\arguments{
\item{x}{LingamResult object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for LingamResult
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()
result <- lingam_direct(LiNGAM_sample_1000$data, reg_method = "ols")
print(result)
}
