% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_multi_group_bootstrap.r
\name{print.MultiGroupBootstrapResult}
\alias{print.MultiGroupBootstrapResult}
\title{Print method for MultiGroupBootstrapResult}
\usage{
\method{print}{MultiGroupBootstrapResult}(x, ...)
}
\arguments{
\item{x}{MultiGroupBootstrapResult object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for MultiGroupBootstrapResult
}
\examples{
mg <- generate_multi_group_sample()
bs <- lingam_multi_group_bootstrap(mg$data_list,
  n_sampling = 10L, reg_method = "ols", seed = 42
)
print(bs)
}
