% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_multi_group.r
\name{print.MultiGroupLingamResult}
\alias{print.MultiGroupLingamResult}
\title{Print method for MultiGroupLingamResult}
\usage{
\method{print}{MultiGroupLingamResult}(x, digits = 3, ...)
}
\arguments{
\item{x}{MultiGroupLingamResult object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for MultiGroupLingamResult
}
\examples{
mg <- generate_multi_group_sample()
res <- lingam_multi_group(mg$data_list, reg_method = "ols")
print(res)
}
