% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_var_bootstrap.r
\name{print.VARBootstrapResult}
\alias{print.VARBootstrapResult}
\title{Print a VARBootstrapResult}
\usage{
\method{print}{VARBootstrapResult}(x, ...)
}
\arguments{
\item{x}{a VARBootstrapResult object}

\item{...}{additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print a VARBootstrapResult
}
\examples{
s <- generate_varlingam_sample(n = 500, seed = 42)
bs <- lingam_var_bootstrap(s$data,
  n_sampling = 10L, lags = 1, criterion = NULL,
  reg_method = "ols", prune = FALSE, seed = 1, verbose = FALSE
)
print(bs)
}
