% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_varma_bootstrap.r
\name{print.VARMABootstrapResult}
\alias{print.VARMABootstrapResult}
\title{Print a VARMABootstrapResult}
\usage{
\method{print}{VARMABootstrapResult}(x, ...)
}
\arguments{
\item{x}{a VARMABootstrapResult object}

\item{...}{additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print a VARMABootstrapResult
}
\examples{
s <- generate_varmalingam_sample(n = 300, seed = 42)
bs <- lingam_varma_bootstrap(s$data,
  n_sampling = 5L, order = c(1, 1), criterion = NULL,
  reg_method = "ols", prune = FALSE, seed = 1, verbose = FALSE
)
print(bs)
}
