% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_rcd.r
\name{rcd_is_correlated}
\alias{rcd_is_correlated}
\title{Pearson correlation test used throughout RCD}
\usage{
rcd_is_correlated(a, b, cor_alpha)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}

\item{cor_alpha}{significance level}
}
\value{
TRUE if correlated (p-value < cor_alpha)
}
\description{
Pearson correlation test used throughout RCD
}
\keyword{internal}
