% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_rcd.r
\name{rcd_is_independent}
\alias{rcd_is_independent}
\title{Independence judgment used throughout RCD (hsic or fcorr)}
\usage{
rcd_is_independent(a, b, independence, ind_alpha, ind_corr)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}

\item{independence}{"hsic" or "fcorr"}

\item{ind_alpha}{significance level (hsic only)}

\item{ind_corr}{rejection threshold (fcorr only)}
}
\value{
TRUE if independent
}
\description{
Independence judgment used throughout RCD (hsic or fcorr)
}
\keyword{internal}
