% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_rcd.r
\name{rcd_ols_resid_coef}
\alias{rcd_ols_resid_coef}
\title{OLS residual and coefficients (with intercept)}
\usage{
rcd_ols_resid_coef(y, X_exog)
}
\arguments{
\item{y}{response vector}

\item{X_exog}{explanatory-variable matrix (may have 0 columns)}
}
\value{
list(resid = residual vector, coef = coefficient vector excluding intercept)
}
\description{
OLS residual and coefficients (with intercept)
}
\keyword{internal}
