% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{search_candidate}
\alias{search_candidate}
\title{Search for candidate variables}
\usage{
search_candidate(U, Aknw, apply_prior_knowledge_softly, partial_orders)
}
\arguments{
\item{U}{Set of currently undetermined variables}

\item{Aknw}{Prior knowledge matrix}

\item{apply_prior_knowledge_softly}{Whether to apply prior knowledge softly}

\item{partial_orders}{Extracted partial orders}
}
\value{
list(Uc, Vj)
}
\description{
Search for candidate variables
}
\keyword{internal}
