% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{search_causal_order_kernel}
\alias{search_causal_order_kernel}
\title{Causal order search via the kernel method}
\usage{
search_causal_order_kernel(X, U, Uc, Vj)
}
\arguments{
\item{X}{Data matrix}

\item{U}{All variables}

\item{Uc}{Candidate variables}

\item{Vj}{Prior knowledge set}
}
\value{
Index of the selected variable
}
\description{
Causal order search via the kernel method
}
\keyword{internal}
