% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{search_causal_order_pwling}
\alias{search_causal_order_pwling}
\title{Causal order search via pwling}
\usage{
search_causal_order_pwling(X, U, Uc, Vj)
}
\arguments{
\item{X}{Data matrix}

\item{U}{Indices of all variables}

\item{Uc}{Indices of candidate variables}

\item{Vj}{Variable set based on prior knowledge}
}
\value{
Index of the selected variable
}
\description{
Causal order search via pwling
}
\keyword{internal}
