% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{skewness_pop}
\alias{skewness_pop}
\title{Skewness (divided by n)}
\usage{
skewness_pop(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Skewness (divided by n)
}
\keyword{internal}
