% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lingam_sample.r
\name{true_adjacency_matrix_10}
\alias{true_adjacency_matrix_10}
\title{True adjacency matrix for 10-variable sample}
\usage{
true_adjacency_matrix_10(labels = paste0("x", 0:9))
}
\arguments{
\item{labels}{variable names (default: x0-x9)}
}
\value{
true adjacency matrix (10 x 10)
}
\description{
True adjacency matrix for 10-variable sample
}
