% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_with_imputation.r
\name{validate_cd_fit_output}
\alias{validate_cd_fit_output}
\title{Validate the return value of a \code{cd_fit} function}
\usage{
validate_cd_fit_output(cd_res, p, n_datasets)
}
\arguments{
\item{cd_res}{Return value of \code{cd_fit(X_list)}}

\item{p}{Number of features}

\item{n_datasets}{Expected number of adjacency matrices (= number of imputed datasets)}
}
\description{
Validate the return value of a \code{cd_fit} function
}
\keyword{internal}
