% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_with_imputation.r
\name{validate_imputer_output}
\alias{validate_imputer_output}
\title{Validate the return value of an \code{imputer} function}
\usage{
validate_imputer_output(datasets, X_boot)
}
\arguments{
\item{datasets}{Return value of \code{imputer(X_boot)}}

\item{X_boot}{The bootstrap-resampled data passed to \code{imputer}}
}
\description{
Validate the return value of an \code{imputer} function
}
\keyword{internal}
