% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{validate_lingam_result}
\alias{validate_lingam_result}
\title{Validate the return value of lingam_direct()}
\usage{
validate_lingam_result(x)
}
\description{
Validate the return value of lingam_direct()
}
\keyword{internal}
