% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_parce.r
\name{validate_parce_result}
\alias{validate_parce_result}
\title{Validate the return value of lingam_parce()}
\usage{
validate_parce_result(x)
}
\description{
Validate the return value of lingam_parce()
}
\keyword{internal}
