% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_rcd.r
\name{validate_rcd_result}
\alias{validate_rcd_result}
\title{Validate the return value of lingam_rcd()}
\usage{
validate_rcd_result(x)
}
\description{
Validate the return value of lingam_rcd()
}
\keyword{internal}
