% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_varma_total_effect.r
\name{varma_joined_design}
\alias{varma_joined_design}
\title{Build the joined lagged design for \code{\link[=varma_total_effect_core]{varma_total_effect_core()}}}
\usage{
varma_joined_design(X, ee_full, order, from_lag)
}
\arguments{
\item{X}{numeric matrix (n x m), rows ordered in time}

\item{ee_full}{LiNGAM residuals, full length (n x m)}

\item{order}{VARMA order c(p, q)}

\item{from_lag}{lag of the source variable (non-negative integer)}
}
\value{
matrix (n x m(1 + p + q + 2*from_lag))
}
\description{
\verb{[X_t, ..., X_\{t-(p+from_lag)\}, e_\{t-1\}, ..., e_\{t-(q+from_lag)\}]}.
}
\keyword{internal}
