% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{calculate_total_effect}
\alias{calculate_total_effect}
\title{総合因果効果を計算}
\usage{
calculate_total_effect(adjacency_matrix, from_index, to_index)
}
\arguments{
\item{adjacency_matrix}{隣接行列}

\item{from_index}{原因変数のインデックス (1-based)}

\item{to_index}{結果変数のインデックス (1-based)}
}
\value{
総合因果効果
}
\description{
総合因果効果を計算
}
\keyword{internal}
