% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{create_bootstrap_result}
\alias{create_bootstrap_result}
\title{BootstrapResult を作成}
\usage{
create_bootstrap_result(
  adjacency_matrices,
  total_effects,
  resampled_indices = NULL
)
}
\arguments{
\item{adjacency_matrices}{array (n_sampling x n_features x n_features)}

\item{total_effects}{array (n_sampling x n_features x n_features)}

\item{resampled_indices}{list of index vectors}
}
\value{
BootstrapResult (list with class attribute)
}
\description{
BootstrapResult を作成
}
\keyword{internal}
