% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{diff_mutual_info}
\alias{diff_mutual_info}
\title{相互情報量の差}
\usage{
diff_mutual_info(xi_std, xj_std, ri_j, rj_i)
}
\arguments{
\item{xi_std}{標準化された xi}

\item{xj_std}{標準化された xj}

\item{ri_j}{xi を xj で回帰した残差}

\item{rj_i}{xj を xi で回帰した残差}
}
\value{
相互情報量の差
}
\description{
相互情報量の差
}
\keyword{internal}
