% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_total_effect.r
\name{estimate_all_total_effects}
\alias{estimate_all_total_effects}
\title{全変数間の総合因果効果を一括推定}
\usage{
estimate_all_total_effects(X, lingam_result, method = "lasso", lambda = "AIC")
}
\arguments{
\item{X}{元データ (n_samples x n_features)}

\item{lingam_result}{direct_lingam() の返り値}

\item{method}{回帰手法 ("ols", "lasso", "adaptive_lasso")}

\item{lambda}{ラムダ選択 ("lambda.min", "lambda.1se", "AIC", "BIC")}
}
\value{
総合因果効果の行列 (行: 結果変数, 列: 原因変数)
}
\description{
全変数間の総合因果効果を一括推定
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

model <- LiNGAM_sample_1000$data |>
  direct_lingam()

LiNGAM_sample_1000$data |>
  estimate_all_total_effects(model)
}
