% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_total_effect.r
\name{estimate_total_effect}
\alias{estimate_total_effect}
\title{指定した2変数間の総合因果効果を推定}
\usage{
estimate_total_effect(
  X,
  lingam_result,
  from_index,
  to_index,
  method = "adaptive_lasso",
  lambda = "BIC"
)
}
\arguments{
\item{X}{元データ (matrix or data.frame)}

\item{lingam_result}{direct_lingam() の返り値}

\item{from_index}{原因変数 (1-based index or 変数名)}

\item{to_index}{結果変数 (1-based index or 変数名)}

\item{method}{回帰手法 ("ols", "lasso", "adaptive_lasso")デフォルトはadaptive_lasso}

\item{lambda}{ラムダ選択 ("lambda.min", "lambda.1se", "AIC", "BIC", "oracle")デフォルトはBIC}
}
\value{
推定された総合因果効果
}
\description{
指定した2変数間の総合因果効果を推定
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

model <- LiNGAM_sample_1000$data |>
  direct_lingam()

LiNGAM_sample_1000$data |>
  estimate_total_effect(model, 4, 1)
}
