% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{extract_partial_orders}
\alias{extract_partial_orders}
\title{事前知識から部分順序を抽出}
\usage{
extract_partial_orders(pk)
}
\arguments{
\item{pk}{事前知識行列 (NaN = 不明)}
}
\value{
matrix (n x 2), 各行は [from, to] の部分順序
}
\description{
事前知識から部分順序を抽出
}
\keyword{internal}
