% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{fit_adaptive_lasso}
\alias{fit_adaptive_lasso}
\title{Adaptive LASSO 回帰（情報量基準 or CV でラムダ選択）}
\usage{
fit_adaptive_lasso(y, Xp, lambda = "AICc", gamma_weight = 1)
}
\arguments{
\item{y}{目的変数}

\item{Xp}{説明変数行列}

\item{lambda}{ラムダ選択方法 ("lambda.min", "lambda.1se", "AICc", "BIC")}

\item{gamma_weight}{重みの指数}
}
\value{
係数ベクトル
}
\description{
Adaptive LASSO 回帰（情報量基準 or CV でラムダ選択）
}
\keyword{internal}
