% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{fit_lasso}
\alias{fit_lasso}
\title{LASSO 回帰（情報量基準 or CV でラムダ選択）}
\usage{
fit_lasso(y, Xp, lambda = "BIC")
}
\arguments{
\item{y}{目的変数}

\item{Xp}{説明変数行列}

\item{lambda}{ラムダ選択方法
"lambda.min" : CV予測誤差最小
"lambda.1se" : CV 1SEルール
"AIC"       : AIC最小
"BIC"        : BIC最小。デフォルト}
}
\value{
係数ベクトル
}
\description{
LASSO 回帰（情報量基準 or CV でラムダ選択）
}
\keyword{internal}
