% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{get_directed_acyclic_graph_counts}
\alias{get_directed_acyclic_graph_counts}
\title{Get DAG counts}
\usage{
get_directed_acyclic_graph_counts(
  result,
  n_dags = NULL,
  min_causal_effect = NULL,
  split_by_causal_effect_sign = FALSE
)
}
\arguments{
\item{result}{BootstrapResult object}

\item{n_dags}{How many of the top entries to return (NULL = all)}

\item{min_causal_effect}{Minimum threshold for the causal effect (NULL = 0)}

\item{split_by_causal_effect_sign}{Whether to split by the sign of the causal effect}
}
\value{
list(dag = list of data.frames, count = integer vector)
}
\description{
Get DAG counts
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data, n_sampling = 30L, seed = 42)

get_directed_acyclic_graph_counts(bs_model)
}
