% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{get_directed_acyclic_graph_counts}
\alias{get_directed_acyclic_graph_counts}
\title{DAG カウントを取得}
\usage{
get_directed_acyclic_graph_counts(
  result,
  n_dags = NULL,
  min_causal_effect = NULL,
  split_by_causal_effect_sign = FALSE
)
}
\arguments{
\item{result}{BootstrapResult オブジェクト}

\item{n_dags}{上位何件を返すか (NULL = 全て)}

\item{min_causal_effect}{因果効果の最小閾値 (NULL = 0)}

\item{split_by_causal_effect_sign}{因果効果の符号で分割するか}
}
\value{
list(dag = list of data.frames, count = integer vector)
}
\description{
DAG カウントを取得
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- LiNGAM_sample_1000$data |>
  lingam_direct_bootstrap(n_sampling = 30L, seed = 42)

bs_model |>
  get_directed_acyclic_graph_counts()
}
