% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{get_probabilities}
\alias{get_probabilities}
\title{Get bootstrap probabilities}
\usage{
get_probabilities(result, min_causal_effect = NULL)
}
\arguments{
\item{result}{BootstrapResult object}

\item{min_causal_effect}{Minimum threshold for the causal effect (NULL = 0)}
}
\value{
Probability matrix (n_features x n_features)
}
\description{
Get bootstrap probabilities
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data, n_sampling = 30L, seed = 42)

get_probabilities(bs_model)
}
