% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{get_probabilities}
\alias{get_probabilities}
\title{ブートストラップ確率を取得}
\usage{
get_probabilities(result, min_causal_effect = NULL)
}
\arguments{
\item{result}{BootstrapResult オブジェクト}

\item{min_causal_effect}{因果効果の最小閾値 (NULL = 0)}
}
\value{
確率行列 (n_features x n_features)
}
\description{
ブートストラップ確率を取得
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- LiNGAM_sample_1000 |>
  bootstrap_lingam(n_sampling = 30L, seed = 42)

bs_model |>
  get_probabilities()
}
