% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{get_total_causal_effects}
\alias{get_total_causal_effects}
\title{総合因果効果リストを取得}
\usage{
get_total_causal_effects(result, min_causal_effect = NULL)
}
\arguments{
\item{result}{BootstrapResult オブジェクト}

\item{min_causal_effect}{因果効果の最小閾値 (NULL = 0)}
}
\value{
data.frame (from, to, effect, probability)
}
\description{
総合因果効果リストを取得
}
