% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{kernel_mi_core}
\alias{kernel_mi_core}
\title{Kernel-based mutual information: core}
\usage{
kernel_mi_core(E1, x2, kappa, sigma)
}
\arguments{
\item{E1}{Variable-1 matrix precomputed by \code{kernel_mi_prepare()}}

\item{x2}{Vector of variable 2}

\item{kappa}{Regularization parameter}

\item{sigma}{Width of the Gaussian kernel}
}
\value{
Mutual information
}
\description{
The target quantity is the difference of logdets of 2n x 2n matrices, but via
the block structure and the Schur complement it can be computed equivalently
using only an n x n Cholesky decomposition:
\verb{MI = -1/2 * (logdet(tmp2^2 - K2 K1 tmp1^-2 K1 K2) - logdet(tmp2^2))}
}
\keyword{internal}
