% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{kurtosis_pop}
\alias{kurtosis_pop}
\title{Kurtosis (divided by n; excess kurtosis, which is 0 for a normal distribution)}
\usage{
kurtosis_pop(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Kurtosis (divided by n; excess kurtosis, which is 0 for a normal distribution)
}
\keyword{internal}
