% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_lim.r
\name{lim_bic_loss}
\alias{lim_bic_loss}
\title{BIC-penalized DAG score (sign-flipped, i.e. a "loss") used during the local
search phase. \code{W} is interpreted as a 0/1 skeleton in the i -> j
orientation (parents of j are the nonzero rows of column j).}
\usage{
lim_bic_loss(W, X, is_continuous)
}
\description{
BIC-penalized DAG score (sign-flipped, i.e. a "loss") used during the local
search phase. \code{W} is interpreted as a 0/1 skeleton in the i -> j
orientation (parents of j are the nonzero rows of column j).
}
\keyword{internal}
