% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingamr-package.R
\docType{package}
\name{lingamr-package}
\alias{lingamr}
\alias{lingamr-package}
\title{lingamr: LiNGAM Algorithms for Causal Discovery}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R implementation of LiNGAM (Linear Non-Gaussian Acyclic Model) algorithms for causal discovery. Based on the Python implementation by Ikeuchi et al. (2023) \url{https://github.com/cdt15/lingam}. The VAR-LiNGAM residual diagnostics are inspired by the VARLiNGAM R code of Moneta et al. \url{https://sites.google.com/site/dorisentner/publications/VARLiNGAM}.
}
\details{
tutorial: \code{vignette("lingamr", package = "lingamr")}
}
\references{
Shimizu, S., et al. (2011). DirectLiNGAM: A direct method for learning a
linear non-Gaussian structural equation model. \emph{Journal of Machine Learning
Research}, 12, 1225-1248.

Hyvärinen, A., Zhang, K., Shimizu, S., & Hoyer, P. O. (2010). Estimation of a
structural vector autoregression model using non-Gaussianity. \emph{Journal of
Machine Learning Research}, 11, 1709-1731.

Moneta, A., Entner, D., Hoyer, P. O., & Coad, A. (2013). Causal inference by
independent component analysis: Theory and applications. \emph{Oxford Bulletin of
Economics and Statistics}, 75(5), 705-730. VARLiNGAM R code:
\url{https://sites.google.com/site/dorisentner/publications/VARLiNGAM}

Python implementation (DirectLiNGAM, VAR-LiNGAM): cdt15/lingam,
\url{https://github.com/cdt15/lingam}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/morimotoosamu/lingamr}
  \item Report bugs at \url{https://github.com/morimotoosamu/lingamr/issues}
}

}
\author{
\strong{Maintainer}: Osamu Morimoto \email{galactic.supermarket@gmail.com}

Authors:
\itemize{
  \item Osamu Morimoto \email{galactic.supermarket@gmail.com}
}

Other contributors:
\itemize{
  \item T. Ikeuchi [copyright holder]
  \item G. Haraoka [copyright holder]
  \item M. Ide [copyright holder]
  \item W. Kurebayashi [copyright holder]
  \item S. Shimizu [copyright holder]
}

}
\keyword{internal}
