% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingamr-package.R
\docType{package}
\name{lingamr-package}
\alias{lingamr}
\alias{lingamr-package}
\title{lingamr: LiNGAM Algorithms for Causal Discovery}
\description{
R implementation of LiNGAM (Linear Non-Gaussian Acyclic Model) algorithms for causal discovery. Based on the Python implementation by Ikeuchi et al. (2023) \url{https://github.com/cdt15/lingam}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/morimotoosamu/lingamr}
  \item \url{https://morimotoosamu.github.io/lingamr/}
  \item Report bugs at \url{https://github.com/morimotoosamu/lingamr/issues}
}

}
\author{
\strong{Maintainer}: Osamu Morimoto \email{galactic.supermarket@gmail.com}

Authors:
\itemize{
  \item Osamu Morimoto \email{galactic.supermarket@gmail.com}
}

}
\keyword{internal}
