% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{mutual_information_kernel}
\alias{mutual_information_kernel}
\title{Kernel-based mutual information}
\usage{
mutual_information_kernel(x1, x2, param)
}
\arguments{
\item{x1}{Variable 1}

\item{x2}{Variable 2}

\item{param}{Parameter vector (kappa, sigma)}
}
\value{
Mutual information
}
\description{
Kernel-based mutual information
}
\keyword{internal}
