% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{mutual_information_kernel}
\alias{mutual_information_kernel}
\title{カーネル法による相互情報量}
\usage{
mutual_information_kernel(x1, x2, param)
}
\arguments{
\item{x1}{変数1}

\item{x2}{変数2}

\item{param}{パラメータベクトル (kappa, sigma)}
}
\value{
相互情報量
}
\description{
カーネル法による相互情報量
}
\keyword{internal}
