% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{plot_bootstrap_probabilities}
\alias{plot_bootstrap_probabilities}
\title{Draw bootstrap probabilities with DiagrammeR}
\usage{
plot_bootstrap_probabilities(
  result,
  labels = NULL,
  min_causal_effect = NULL,
  min_probability = 0.5,
  rankdir = "TB",
  shape = "circle"
)
}
\arguments{
\item{result}{BootstrapResult object}

\item{labels}{Vector of variable names (NULL allowed)}

\item{min_causal_effect}{Minimum causal effect to display}

\item{min_probability}{Minimum probability to display}

\item{rankdir}{Layout direction}

\item{shape}{Node shape}
}
\value{
grViz object
}
\description{
Draw bootstrap probabilities with DiagrammeR
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data, n_sampling = 30L, seed = 42)
plot_bootstrap_probabilities(bs_model)
}
