% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{plot_bootstrap_probabilities}
\alias{plot_bootstrap_probabilities}
\title{ブートストラップ確率を DiagrammeR で描画}
\usage{
plot_bootstrap_probabilities(
  result,
  labels = NULL,
  min_causal_effect = NULL,
  min_probability = 0.5,
  rankdir = "TB",
  shape = "circle"
)
}
\arguments{
\item{result}{BootstrapResult オブジェクト}

\item{labels}{変数名ベクトル (NULL可)}

\item{min_causal_effect}{表示する最小因果効果}

\item{min_probability}{表示する最小確率}

\item{rankdir}{レイアウト方向}

\item{shape}{ノード形状}
}
\value{
grViz オブジェクト
}
\description{
ブートストラップ確率を DiagrammeR で描画
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- LiNGAM_sample_1000 |>
  bootstrap_lingam(n_sampling = 30L, seed = 42)
bs_model |>
  plot_bootstrap_probabilities()
}
