% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{plot_residual_qq}
\alias{plot_residual_qq}
\title{plot QQ}
\usage{
plot_residual_qq(X, lingam_result, ncol = 3, nrow = NULL)
}
\arguments{
\item{X}{original data (matrix or data.frame)}

\item{lingam_result}{return value of lingam_direct()}

\item{ncol}{Number of columns.}

\item{nrow}{Number of rows.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object with QQ plots of residuals.
}
\description{
plot QQ
}
\examples{
# Load the sample data
LiNGAM_sample_1000 <- generate_lingam_sample_6()

# Run Direct LiNGAM
result <- lingam_direct(LiNGAM_sample_1000$data)

plot_residual_qq(LiNGAM_sample_1000$data, result)
}
