% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{plot_residual_qq}
\alias{plot_residual_qq}
\title{plot QQ}
\usage{
plot_residual_qq(X, lingam_result, ncol = 3, nrow = NULL)
}
\arguments{
\item{X}{元データ (matrix or data.frame)}

\item{lingam_result}{direct_lingam() の返り値}

\item{ncol}{Number of columns.}

\item{nrow}{Number of rows.}
}
\description{
plot QQ
}
\examples{
# サンプルデータの呼び出し
LiNGAM_sample_1000 <- generate_lingam_sample_6()

# Direct LiNGAM の実行
result <- direct_lingam(LiNGAM_sample_1000)

plot_residual_qq(LiNGAM_sample_1000, result)
}
