% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_bootstrap.r
\name{print.BootstrapResult}
\alias{print.BootstrapResult}
\title{BootstrapResult の内容を表示}
\usage{
\method{print}{BootstrapResult}(x, ...)
}
\arguments{
\item{x}{BootstrapResult オブジェクト}

\item{...}{追加の引数 (S3メソッド互換用)}
}
\description{
BootstrapResult の内容を表示
}
\examples{
LiNGAM_sample_1000 <- generate_lingam_sample_6()

bs_model <- lingam_direct_bootstrap(LiNGAM_sample_1000$data, n_sampling = 30L, seed = 42)

print(bs_model)
}
