% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{print.BootstrapResult}
\alias{print.BootstrapResult}
\title{BootstrapResult の内容を表示}
\usage{
\method{print}{BootstrapResult}(x, ...)
}
\arguments{
\item{x}{BootstrapResult オブジェクト}

\item{...}{追加の引数 (S3メソッド互換用)}
}
\description{
BootstrapResult の内容を表示
}
\examples{
data(LiNGAM_sample_1000)

bs_model <- LiNGAM_sample_1000 |>
  bootstrap_lingam(n_sampling = 30L, seed = 42)

bs_model |>
  print()
}
