% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_order_stability.r
\name{print.causal_order_stability}
\alias{print.causal_order_stability}
\title{print method for causal_order_stability}
\usage{
\method{print}{causal_order_stability}(x, ...)
}
\arguments{
\item{x}{A \code{causal_order_stability} object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly.
}
\description{
print method for causal_order_stability
}
