% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_order_stability.r
\name{print.causal_order_stability}
\alias{print.causal_order_stability}
\title{causal_order_stability の print メソッド}
\usage{
\method{print}{causal_order_stability}(x, ...)
}
\arguments{
\item{x}{\code{causal_order_stability} オブジェクト}

\item{...}{追加引数（未使用）}
}
\description{
causal_order_stability の print メソッド
}
