% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_independence_p_values.r
\name{print.lingam_normality_test}
\alias{print.lingam_normality_test}
\title{Print method for lingam_normality_test}
\usage{
\method{print}{lingam_normality_test}(x, ...)
}
\arguments{
\item{x}{lingam_normality_test object}

\item{...}{additional arguments}
}
\description{
Print method for lingam_normality_test
}
