% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lingam.r
\name{print.lingam_summary}
\alias{print.lingam_summary}
\title{lingam_summary の print メソッド}
\usage{
\method{print}{lingam_summary}(x, ...)
}
\arguments{
\item{x}{\code{lingam_summary} オブジェクト}

\item{...}{追加引数（未使用）}
}
\description{
lingam_summary の print メソッド
}
