% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{print.LingamResult}
\alias{print.LingamResult}
\title{LingamResult の print メソッド}
\usage{
\method{print}{LingamResult}(x, digits = 3, ...)
}
\arguments{
\item{x}{LingamResult オブジェクト}

\item{digits}{表示桁数}

\item{...}{追加引数（未使用）}
}
\description{
LingamResult の print メソッド
}
