% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_var_diagnostics.r
\name{print.var_stationarity}
\alias{print.var_stationarity}
\title{Print method for var_stationarity}
\usage{
\method{print}{var_stationarity}(x, ...)
}
\arguments{
\item{x}{a \code{var_stationarity} object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for var_stationarity
}
