% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_var_bootstrap.r
\name{print.VARBootstrapResult}
\alias{print.VARBootstrapResult}
\title{Print a VARBootstrapResult}
\usage{
\method{print}{VARBootstrapResult}(x, ...)
}
\arguments{
\item{x}{a VARBootstrapResult object}

\item{...}{additional arguments (unused)}
}
\description{
Print a VARBootstrapResult
}
