% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_var.r
\name{print.VARLiNGAMResult}
\alias{print.VARLiNGAMResult}
\title{Print method for VARLiNGAMResult}
\usage{
\method{print}{VARLiNGAMResult}(x, digits = 3, ...)
}
\arguments{
\item{x}{VARLiNGAMResult object}

\item{digits}{number of digits to display}

\item{...}{additional arguments (unused)}
}
\description{
Print method for VARLiNGAMResult
}
