% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_causal_order.r
\name{residual_vec}
\alias{residual_vec}
\title{Residual (residual when xi is regressed on xj)
Compute the residual vector}
\usage{
residual_vec(xi, xj, standardized = FALSE)
}
\arguments{
\item{xi}{Target variable vector}

\item{xj}{Explanatory variable vector}

\item{standardized}{Whether the data is already standardized (default: FALSE)}
}
\value{
Residual vector after regression
}
\description{
Residual (residual when xi is regressed on xj)
Compute the residual vector
}
\keyword{internal}
