% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{residual_vec}
\alias{residual_vec}
\title{残差 (xi を xj に回帰したときの残差)
残差ベクトルの計算}
\usage{
residual_vec(xi, xj)
}
\arguments{
\item{xi}{対象変数ベクトル}

\item{xj}{説明変数ベクトル}
}
\value{
回帰後の残差ベクトル
}
\description{
残差 (xi を xj に回帰したときの残差)
残差ベクトルの計算
}
\keyword{internal}
