% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{search_candidate}
\alias{search_candidate}
\title{候補変数の探索}
\usage{
search_candidate(U, Aknw, apply_prior_knowledge_softly, partial_orders)
}
\arguments{
\item{U}{現在の未確定変数の集合}

\item{Aknw}{事前知識行列}

\item{apply_prior_knowledge_softly}{ソフト適用の有無}

\item{partial_orders}{抽出された部分順序}
}
\value{
list(Uc, Vj)
}
\description{
候補変数の探索
}
\keyword{internal}
