% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingam_direct.r
\name{search_causal_order_kernel}
\alias{search_causal_order_kernel}
\title{カーネル法による因果順序の探索}
\usage{
search_causal_order_kernel(X, U, Uc, Vj)
}
\arguments{
\item{X}{データ行列}

\item{U}{全変数}

\item{Uc}{候補変数}

\item{Vj}{事前知識集合}
}
\value{
選ばれた変数のインデックス
}
\description{
カーネル法による因果順序の探索
}
\keyword{internal}
