% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_lingam.r
\name{search_causal_order_pwling}
\alias{search_causal_order_pwling}
\title{pwling による因果順序の探索}
\usage{
search_causal_order_pwling(X, U, Uc, Vj)
}
\arguments{
\item{X}{データ行列}

\item{U}{全変数インデックス}

\item{Uc}{候補変数のインデックス}

\item{Vj}{事前知識に基づく変数集合}
}
\value{
選ばれた変数のインデックス
}
\description{
pwling による因果順序の探索
}
\keyword{internal}
